feat: add agreement confirmation before user delte entire tree

+++ src/Window/OptionsFrontend.js
@@ -63,6 +63,10 @@ class Options extends OptionsEngine {
   }
 
   async deleteTree() {
+    const userAgreed = confirm("Do you want to delete this tree ?")
+
+    if (!userAgreed) return
+    
     await super.deleteTreeFromDisc();
     super.reload();
   }
